requirements.txt
Flask==3.0.0
Flask-SocketIO==5.3.6
google-generativeai==0.7.2
twilio==8.10.0
SpeechRecognition==3.10.0
gTTS==2.4.0
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
python-socketio==5.9.0
eventlet==0.33.3